🎨 Improve code structure
